+++ tools/rendercli/rendercli.cpp
@@ -170,6 +170,7 @@ namespace {
         << " integrator=" << batching.value("integrator").toString().toStdString()
         << " execution=" << batching.value("executionMode").toString().toStdString()
         << " samples=" << unsignedValue(input, "primarySamples")
+        << " active_sample_depths=" << unsignedValue(batching, "activeSampleDepthsProcessed")
         << " batches=" << unsignedValue(batching, "batches")
         << " avg_batch=" << batching.value("averageBatchSize").toDouble()
         << " max_batch=" << unsignedValue(batching, "maxBatchSize")